refactor(framework): extract the transaction cost model into TxGasModel.h
Admission and execution must compute intrinsic gas from the same formula. Today
the only implementation over protocol::Transaction lives inside
ethereum-executor/EthereumTransition.h, which no admission path can reach:
including it drags in EthereumHost.h / EthereumState.h / evmone, and neither
txpool nor rpc links evmone. Copying the formula instead would drift at the next
hard fork that moves EIP-7623's min_cost or the EIP-7702 authorization cost --
and a drift shows up as "admission accepted, execution returned OutOfGasLimit",
i.e. a block carrying a deterministically failing transaction.

New header bcos-framework/bcos-framework/protocol/TxGasModel.h holds, in
namespace bcos::protocol:

  ethToAddress, num_words, compute_tx_data_tokens, compute_access_list_cost,
  TransactionCost, compute_tx_intrinsic_cost   (from EthereumTransition.h)
  MAX_TX_GAS_LIMIT, AUTHORIZATION_EMPTY_ACCOUNT_COST, toIntxU256
                                               (from EVMSupport.h)

Its only dependencies are the header-only evmc and intx headers plus
protocol/Transaction.h. bcos-framework now PUBLIC-links intx::intx under
FULLNODE -- intx is an INTERFACE target, so this adds include directories and no
archive to any consumer's link line, unlike evmone, which is deliberately still
includes-only there.

Both former homes keep their namespaces and re-import the symbols with
using-declarations, so all existing call sites (~20 uses of evm::toIntxU256
across ethereum-executor and opstack-executor, and the qualified
eth_transition_detail::compute_tx_intrinsic_cost uses) are untouched. No
behaviour change.

Verified:
  cmake --build build --target ethereum-executor opstack-executor
      test-transaction-scheduler test-ethereum-state-smoke  -> all built clean
  ./build/ethereum-executor/tests/test-ethereum-state-smoke -> all checks passed
  ./build/transaction-scheduler/tests/test-transaction-scheduler
      --run_test=TestEthereumExecutorScheduler                -> No errors detected
